tickets/src/events/listeners/order-updated-listener.ts
import { Message } from "node-nats-streaming";
import { queueGrouName } from "./queue-group-name";
import { Listener, OrderCancelledEvent, Subjects } from "@thinkingcorp/common";
import { Ticket } from "../../models/ticket";
import { TicketUpdatedPublisher } from "../publishers/ticket-updated-publisher";

export class OrderCancelledListener extends Listener<OrderCancelledEvent> {
  subject: Subjects.OrderCancelled = Subjects.OrderCancelled;
  queueGrouName = queueGrouName;

  async onMessage(data: OrderCancelledEvent["data"], msg: Message) {
    const ticket = await Ticket.findById(data.ticket.id);

    //if no ticket, throw error
    if (!ticket) {
      throw new Error("Ticket not found");
    }

    //mark ticket as being reserved by settings its orderId property
    ticket.set({ orderId: undefined });

    //save ticket
    await ticket.save();
    await new TicketUpdatedPublisher(this.client).publish({
      id: ticket.id,
      version: ticket.version,
      title: ticket.title,
      price: ticket.price,
      userId: ticket.userId,
      orderId: ticket.orderId,
    });

    //ack message
    msg.ack();
  }
}
